SQL needs to store cbq as a real 0..1

+++ structure.sql
@@ -55,7 +55,7 @@ CREATE TABLE IF NOT EXISTS `users` (
   `session` varchar(40) COLLATE utf8_unicode_ci NOT NULL,
   `repo_name` varchar(64) COLLATE utf8_unicode_ci NOT NULL,
   `com` int(11) NOT NULL,
-  `cbq` int(11) NOT NULL,
+  `cbq` double NOT NULL,
   `issue` int(11) NOT NULL,
   `issue_eta` int(10) NOT NULL,
   UNIQUE KEY `uid` (`uid`),
